fix(auth): bound the notification probe on every sign-in path
The previous commit capped the probe at the OAuth callback only. The Mini App
auth route and the account-link route call the same helper and were left
unbounded, so a slow Bot API could still spend the whole proxy budget there
and answer the customer with a 504.

The cap now lives in the shared helper instead of one call site: a stalled
Bot API costs a bounded wait and an unknown status, which is what the helper
already returns for every other failure, and no future caller can reintroduce
an unbounded probe by forgetting to wrap it.
